mb/google/hatch/var/dratini: Add FP enable
Add FP enable/disable based on SKU ID for Dratini. This is meant
to resolve a UMA issue with Dratini devices that had the FPMCU
populated on non-fp devices.  Since the FPMCU is present, and the
firmware enables the power GPIO's based on variant, not SKU, the
devices were reporting data on fingerprint errantly.

BUG=b:354769653
BUG=b:200825114
TEST=Flash to Dratini, test FP.
Disable test SKU, flash on Dratini, test FP.

To test, run `ectool --name=cros_fp version` in the shell
When enabled, the fpmcu fw version should be displayed.
When disabled, an error should be displayed because the fpmcu
is inaccessible.

Change-Id: Ifc450f51b00b9c3b62268ce94884f5749a3e18c0
Signed-off-by: Jon Murphy <jpmurphy@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/83745
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Martin L Roth <gaumless@gmail.com>
Reviewed-by: Paul Menzel <paulepanter@mailbox.org>