soc/intel/alderlake: Correct ISH partition availability check
The previous implementation incorrectly assumed that the presence of a
UFS device implied the availability of the ISH partition. This is not
always true, especially on Alder Lake platforms where ISH may be
enabled by default even without UFS.

This patch fixes the issue by directly checking for the presence of the
ISH device to determine if the ISH partition is available.

BUG=b:359440547
TEST=1. Able to dump the ISH version with UFS device:

```
tirwen-rev3 ~ # cbmem -c -1 | grep ISH
[DEBUG]  ISH version: 5.4.2.7780
```

2. Able to dump the ISH version with eMMC device:

```
trulo-rev1 ~ # cbmem -c | grep ISH
[DEBUG]  ISH version: 5.4.2.7780
```

Change-Id: I411e36606c0697f91050af40e0636f7c64810e95
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/83898
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Dinesh Gehlot <digehlot@google.com>
Reviewed-by: Eric Lai <ericllai@google.com>